restructure of auth directory
