+++ .betterci/config.json
@@ -12,8 +12,8 @@
     "push":true
   },{
     "name":"auth-server",
-    "file":"litmus-portal/backend/auth/Dockerfile",
-    "context":"litmus-portal/backend/auth",
+    "file":"litmus-portal/authentication/Dockerfile",
+    "context":"litmus-portal/authentication",
     "push":true
   }],
   "k8s-manifest":".betterci/ci-k8s-manifest.yml"
