+++ .circleci/config.yml
@@ -71,7 +71,7 @@ jobs:
       - run:
           name: Build auth server docker image
           command: docker images && docker build . -f Dockerfile -t ${REPONAME}/${AUTHENTICATION_SERVER_IMAGE}:${IMGTAG}
-          working_directory: ~/project/litmus-portal/backend/auth
+          working_directory: ~/project/litmus-portal/authentication
       - run:
           name: Save auth server docker image
           command: docker save -o /tmp/workspace/${AUTHENTICATION_SERVER_IMAGE}.tar ${REPONAME}/${AUTHENTICATION_SERVER_IMAGE}:${IMGTAG}
