+++ litmus-portal/Makefile
@@ -61,9 +61,9 @@ backend-services-checks:
 		&& exit 1; \
 	fi
 	@echo "------------------"
-	@echo "--> Check litmus-portal auth [go mod tidy]"
+	@echo "--> Check litmus-portal authentication [go mod tidy]"
 	@echo "------------------"
-	@tidyRes=$$(cd backend/auth && go mod tidy); \
+	@tidyRes=$$(cd authentication && go mod tidy); \
 	if [ -n "$${tidyRes}" ]; then \
 		echo "go mod tidy checking failed!" && echo "$${tidyRes}" \
 		&& echo "Please ensure you are using $$($(GO) version) for formatting code." \
