+++ litmus-portal/authentication/.gitignore
@@ -1,2 +1,1 @@
 # Build Files
-server
