+++ litmus-portal/authentication/go.mod
@@ -1,0 +1,14 @@
+module github.com/litmuschaos/litmus/litmus-portal/authentication
+
+go 1.13
+
+require (
+	github.com/dgrijalva/jwt-go v3.2.0+incompatible
+	github.com/gin-contrib/cors v1.3.1
+	github.com/gin-gonic/gin v1.6.3
+	github.com/globalsign/mgo v0.0.0-20181015135952-eeefdecb41b8
+	github.com/golang/glog v0.0.0-20160126235308-23def4e6c14b
+	github.com/litmuschaos/litmus/litmus-portal/backend/auth v0.0.0-20200912100951-f970fad0b854 // indirect
+	golang.org/x/crypto v0.0.0-20200622213623-75b288015ac9
+	golang.org/x/oauth2 v0.0.0-20200107190931-bf48bf16ab8d
+)
