+++ litmus-portal/authentication/pkg/types/const.go
